Moved the usecases to a separate package, so the ui package is purely UI only.
