+++ app/src/test/java/com/codemate/koffeemate/usecases/SendCoffeeAnnouncementUseCaseTest.kt
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package com.codemate.koffeemate.ui.main
+package com.codemate.koffeemate.usecases
 
 import com.codemate.koffeemate.BuildConfig
 import com.codemate.koffeemate.data.network.SlackApi
 import com.codemate.koffeemate.data.network.SlackService
-import okhttp3.Dispatcher
 import okhttp3.ResponseBody
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
